refactor: Lotto, WinningLotto 클래스를 model.lotto 패키지로 이동
